move hash to extension
