+++ Extensions/StringExtensions.cs
@@ -1,4 +1,6 @@
 using System.Reflection.Emit;
+using System.Security.Cryptography;
+using System.Text;
 using BaseLib.Utils.NodeFactories;
 using Godot;
 using HarmonyLib;
@@ -38,5 +40,37 @@ public static class StringExtensions
     internal static void WriteLogObj(object? o)
     {
         BaseLibMain.Logger.Info(o?.ToString() ?? "NULL");
+    }
+    
+    private static readonly HashAlgorithm MD5 = System.Security.Cryptography.MD5.Create(); //Not for security, just for comparison.
+    private static readonly Dictionary<string, int> HashDict = [];
+    private static readonly HashSet<int> ExistingHashes = [];
+    
+    public static int ComputeBasicHash(this string s)
+    {
+        if (!HashDict.TryGetValue(s, out var hash))
+        {
+            var data = MD5.ComputeHash(Encoding.UTF8.GetBytes(s));
+            unchecked
+            {
+                const int p = 16777619;
+            
+                hash = (int)2166136261;
+                for (int i = 0; i < data.Length; i++)
+                    hash = (hash ^ data[i]) * p;
+                HashDict[s] = hash;
+                if (ExistingHashes.Add(hash)) return hash;
+                
+                foreach (var entry in HashDict)
+                {
+                    if (entry.Value.Equals(hash))
+                    {
+                        BaseLibMain.Logger.Warn($"Duplicate hashes for {entry.Key} and {s}: {hash}");
+                    }
+                }
+                return hash;
+            }
+        }
+        return hash;
     }
 }
