+++ Patches/Content/CustomEnums.cs
@@ -81,9 +81,6 @@ public static class CustomKeywords
 
 public static class CustomEnums
 {
-    private static readonly HashAlgorithm MD5 = System.Security.Cryptography.MD5.Create(); //Not for security, just for comparison.
-    private static readonly Dictionary<string, int> HashDict = [];
-    private static readonly HashSet<int> ExistingHashes = [];
     private static readonly Dictionary<Type, KeyGenerator> KeyGenerators = [];
     
     /// <summary>All custom enum values generated by BaseLib, grouped by enum type. Maps int value to field name.</summary>
@@ -106,35 +103,7 @@ public static class CustomEnums
             KeyGenerators.Add(enumType, generator = new(enumType));
         }
     
-        return generator.GetKey(ComputeBasicHash(@namespace), ComputeBasicHash(name));
-    }
-
-    private static int ComputeBasicHash(string s)
-    {
-        if (!HashDict.TryGetValue(s, out var hash))
-        {
-            var data = MD5.ComputeHash(Encoding.UTF8.GetBytes(s));
-            unchecked
-            {
-                const int p = 16777619;
-            
-                hash = (int)2166136261;
-                for (int i = 0; i < data.Length; i++)
-                    hash = (hash ^ data[i]) * p;
-                HashDict[s] = hash;
-                if (ExistingHashes.Add(hash)) return hash;
-                
-                foreach (var entry in HashDict)
-                {
-                    if (entry.Value.Equals(hash))
-                    {
-                        BaseLibMain.Logger.Warn($"Duplicate mod hash for {entry.Key} and {s}: {hash}");
-                    }
-                }
-                return hash;
-            }
-        }
-        return hash;
+        return generator.GetKey(@namespace.ComputeBasicHash(), name.ComputeBasicHash());
     }
     
     private class KeyGenerator
